backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.5
qdrant-client==1.7.0
cohere==4.3.1
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
aiohttp==3.9.0
beautifulsoup4==4.12.2
lxml==4.9.3
